refactor: Controller를 Game과 Race로 분리
